# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "独角数卡 API V1",
        "description": "独角数卡项目的API V1版本，提供了RESTful风格的JSON API接口",
        "contact": {
            "name": "独角数卡开发团队",
            "url": "https://utf8.hk",
            "email": "Ashang@utf8.hk"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://unishop.wnecn.com",
            "description": "测试环境"
        }
    ],
    "paths": {
        "/api/v1/goods/list": {
            "get": {
                "tags": [
                    "商品管理"
                ],
                "summary": "获取商品列表",
                "description": "获取所有已开启的商品信息列表",
                "operationId": "getGoodsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "每页数量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "商品分类ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "商品类型 1:自动发货 2:人工处理",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "搜索关键词",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "required": [
                                                "data"
                                            ],
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/GoodsListData"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/goods/category/list": {
            "get": {
                "tags": [
                    "商品管理"
                ],
                "summary": "获取商品分类",
                "description": "获取所有商品分类信息",
                "operationId": "getGoodsCategories",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "required": [
                                                "data"
                                            ],
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/CategoryListData"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/create_order": {
            "post": {
                "tags": [
                    "订单-游客模式"
                ],
                "summary": "创建订单",
                "description": "创建新的商品订单（非登录状态）",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            },
                            "examples": {
                                "example1": {
                                    "summary": "创建订单示例",
                                    "value": {
                                        "gid": 1,
                                        "email": "testuser@abc.com",
                                        "payway": 1,
                                        "buy_amount": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "required": [
                                                "data"
                                            ],
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/CreateOrderData"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Goods": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "gd_name": {
                        "type": "string",
                        "example": "测试商品"
                    },
                    "gd_description": {
                        "type": "string",
                        "example": "商品描述"
                    },
                    "gd_keywords": {
                        "type": "string",
                        "example": "关键词"
                    },
                    "picture": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "retail_price": {
                        "type": "string",
                        "example": "100.00"
                    },
                    "actual_price": {
                        "type": "string",
                        "example": "80.00"
                    },
                    "in_stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "sales_volume": {
                        "type": "integer",
                        "example": 50
                    },
                    "type": {
                        "description": "1:自动发货 2:人工处理",
                        "type": "integer",
                        "example": 1
                    },
                    "type_text": {
                        "type": "string",
                        "example": "自动发货"
                    },
                    "is_open": {
                        "description": "1:开启 0:关闭",
                        "type": "integer",
                        "example": 1
                    },
                    "is_open_text": {
                        "type": "string",
                        "example": "开启"
                    },
                    "buy_limit_num": {
                        "type": "integer",
                        "example": 10
                    },
                    "buy_prompt": {
                        "type": "string",
                        "example": "购买提示"
                    },
                    "description": {
                        "type": "string",
                        "example": "商品详细描述"
                    },
                    "ord": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "group": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "gp_name": {
                                "type": "string",
                                "example": "商品分类"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "GoodsListData": {
                "properties": {
                    "goods": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Goods"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/PaginationResponse"
                    }
                },
                "type": "object"
            },
...
            "GoodsGroup": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "gp_name": {
                        "type": "string",
                        "example": "商品分类"
                    },
                    "ord": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_open": {
                        "description": "1:开启 0:关闭",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "CategoryListData": {
                "properties": {
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GoodsGroup"
                        }
                    }
                },
                "type": "object"
            },
            "PaginationResponse": {
                "properties": {
                    "current_page": {
                        "description": "当前页码",
                        "type": "integer",
                        "example": "1"
                    },
                    "per_page": {
                        "description": "每页显示数量",
                        "type": "integer",
                        "example": "20"
                    },
                    "total": {
                        "description": "总条数",
                        "type": "integer",
                        "example": "21"
                    },
                    "last_page": {
                        "description": "最后一页",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            },
...
            "CreateOrderRequest": {
                "required": [
                    "gid",
                    "email",
                    "payway",
                    "buy_amount"
                ],
                "properties": {
                    "gid": {
                        "description": "商品ID",
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "description": "邮箱地址",
                        "type": "string",
                        "format": "email",
                        "example": "testuser@abc.com"
                    },
                    "payway": {
                        "description": "支付方式编号",
                        "type": "integer",
                        "example": 1
                    },
                    "buy_amount": {
                        "description": "购买数量",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CreateOrderData": {
                "properties": {
                    "order_sn": {
                        "description": "订单号",
                        "type": "string",
                        "example": "ABC123DEF456GHI7"
                    }
                },
                "type": "object"
            },
...
            "BaseResponse": {
                "description": "通用响应结构基类",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "description": "状态码",
                        "type": "integer"
                    },
                    "message": {
                        "description": "响应信息",
                        "type": "string"
                    },
                    "data": {
                        "nullable": true,
                        "description": "响应数据"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Bearer Token认证",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}